Prediction of Annual Prices of Gold
Gold has been a symbol of prestige, power, and prosperity for a very long period of time. It provides unwavering stability as an asset that has historically kept its value in the face of political and economic turmoil. The prediction of gold prices is, therefore, one of the most popular and interesting subjects that analysts research and work on. The prediction happens for a different number of time periods like monthly, weekly, daily etc. This project will utilize an annual data of gold prices to perform time-series analysis and predict the gold
prices for the next 10 years.